Do not re-test unified logging features
